data/requirements.txt
pandas==2.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.19
PyMySQL==1.1.0
cryptography==41.0.3
